Split analysis/neuron.rs into focused sub-modules (#598)
Split the monolithic neuron.rs (~928 lines) into focused sub-modules:
- neuron/mod.rs — Public API, orchestration, parallel loop
- neuron/preparation.rs — Focus target filtering, type maps, source loading
- neuron/evaluation.rs — GPU candidate evaluation (ReLU, activation specs)
- neuron/post_processing.rs — Impact discounting, sorting, result assembly

All files under ~450 lines, well within the ~1,500 line target.
Public API unchanged. All existing tests pass.
